requirements.txt
anyio==2.0.2
asgiref==3.3.4
asyncclick==7.1.2.3
bcrypt==3.2.0
boto3==1.17.109
botocore==1.20.109
certifi==2021.5.30
cffi==1.14.4
charset-normalizer==2.0.1
dj-database-url==0.5.0
Django==3.2.2
django-heroku==0.3.1
django-phone-field==1.8.1
django-storages==1.11.1
gunicorn==20.1.0
idna==3.0
importlib-metadata==3.3.0
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.18.0
python-kasa==0.4.0.dev2
pytz==2020.5
requests==2.26.0
s3transfer==0.4.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.6
whitenoise==5.2.0
xlwt==1.3.0
zipp==3.4.0
